chore: refactor fetching ISC assets
