+++ packages/shared/src/lib/core/layer-2/utils/index.ts
@@ -4,6 +4,7 @@ export * from './buildEvmTransactionDataForNft'
 export * from './buildEvmTransactionDataForToken'
 export * from './calculateEstimatedGasFeeFromTransactionData'
 export * from './calculateMaxGasFeeFromTransactionData'
+export * from './fetchIscAssetsForAccount'
 export * from './fetchL2BalanceForAllAccounts'
 export * from './getAmountFromEvmTransactionValue'
 export * from './getEvmTokenMetadata'
